BugFix: steerable implements were displayed (bump to 1.0.1.5)
Some custom vehicle types that contained the "enterable" specialization, but aren't really drivable vehicles were getting through the filters and stopping the script.  Those are now correctly filtered out, along with some extra logic to handle them gracefully just in case they are not.  Thanks to Yeldawg for finding this.